+++ packages/process/lib/Polysemy/Process.hs
@@ -28,6 +28,7 @@ module Polysemy.Process (
   SystemProcess,
   withSystemProcess,
   withSystemProcess_,
+  Pid (..),
 
   -- ** Pty
   Pty,
@@ -95,12 +96,21 @@ module Polysemy.Process (
 
 import Prelude hiding (send)
 
+import Polysemy.Process.Data.Pid (Pid (..))
 import Polysemy.Process.Data.ProcessError (ProcessError (..))
 import Polysemy.Process.Data.ProcessKill (ProcessKill (..))
 import Polysemy.Process.Data.ProcessOptions (ProcessOptions (ProcessOptions))
 import Polysemy.Process.Data.ProcessOutputParseResult (ProcessOutputParseResult (..))
 import Polysemy.Process.Data.SystemProcessError (SystemProcessError)
-import Polysemy.Process.Effect.Process (Process (..), recv, send, withProcess, withProcessOneshot, withProcessOneshot_, withProcess_)
+import Polysemy.Process.Effect.Process (
+  Process (..),
+  recv,
+  send,
+  withProcess,
+  withProcessOneshot,
+  withProcessOneshot_,
+  withProcess_,
+  )
 import Polysemy.Process.Effect.ProcessInput (ProcessInput)
 import Polysemy.Process.Effect.ProcessOutput (OutputPipe (Stderr, Stdout), ProcessOutput)
 import Polysemy.Process.Effect.Pty (Pty, withPty)